Split out PartType into ShipPart.{h,cpp}
