+++ UI/ClientUI.cpp
@@ -24,6 +24,7 @@
 #include "../universe/System.h"
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
+#include "../universe/ShipPart.h"
 #include "../universe/ShipPartHull.h"
 #include "../universe/Tech.h"
 #include "../universe/Special.h"
