+++ msvc2017/Common/Common.vcxproj.filters
@@ -130,6 +130,9 @@
     <ClInclude Include="..\..\universe\ShipDesign.h">
       <Filter>Header Files\universe</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\universe\ShipPart.h">
+      <Filter>Header Files\universe</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\universe\Special.h">
       <Filter>Header Files\universe</Filter>
     </ClInclude>
@@ -392,6 +395,9 @@
       <Filter>Source Files\universe</Filter>
     </ClCompile>
     <ClCompile Include="..\..\universe\ShipDesign.cpp">
+      <Filter>Source Files\universe</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\universe\ShipPart.cpp">
       <Filter>Source Files\universe</Filter>
     </ClCompile>
     <ClCompile Include="..\..\universe\Special.cpp">
