+++ parse/ShipPartsParser.cpp
@@ -11,7 +11,7 @@
 #include "CommonParamsParser.h"
 
 #include "../universe/Condition.h"
-#include "../universe/ShipPartHull.h"
+#include "../universe/ShipPart.h"
 #include "../universe/ValueRef.h"
 
 #include <boost/spirit/include/phoenix.hpp>
