+++ python/UniverseWrapper.cpp
@@ -4,6 +4,7 @@
 #include "../universe/Fleet.h"
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
+#include "../universe/ShipPart.h"
 #include "../universe/ShipPartHull.h"
 #include "../universe/Building.h"
 #include "../universe/ResourceCenter.h"
