+++ universe/CMakeLists.txt
@@ -22,6 +22,7 @@ target_sources(freeorioncommon
         ${CMAKE_CURRENT_LIST_DIR}/ResourceCenter.h
         ${CMAKE_CURRENT_LIST_DIR}/ScriptingContext.h
         ${CMAKE_CURRENT_LIST_DIR}/ShipDesign.h
+        ${CMAKE_CURRENT_LIST_DIR}/ShipPart.h
         ${CMAKE_CURRENT_LIST_DIR}/ShipPartHull.h
         ${CMAKE_CURRENT_LIST_DIR}/Ship.h
         ${CMAKE_CURRENT_LIST_DIR}/Special.h
@@ -53,6 +54,7 @@ target_sources(freeorioncommon
         ${CMAKE_CURRENT_LIST_DIR}/ResourceCenter.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Ship.cpp
         ${CMAKE_CURRENT_LIST_DIR}/ShipDesign.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/ShipPart.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Special.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Species.cpp
         ${CMAKE_CURRENT_LIST_DIR}/System.cpp
