+++ universe/Conditions.cpp
@@ -11,6 +11,7 @@
 #include "Fleet.h"
 #include "Ship.h"
 #include "ShipDesign.h"
+#include "ShipPart.h"
 #include "ShipPartHull.h"
 #include "ObjectMap.h"
 #include "Planet.h"
